finish line

+++ src/app/models/blog/blog.service.ts
@@ -1,4 +1,3 @@
-import { log } from "console";
 import QueryBuilder from "../../builder/QueryBuilders";
 import { User } from "../user/user.model";
 import { IBlog } from "./blog.interface";
@@ -6,63 +5,61 @@ import { Blog } from "./blog.model";
 import { blogSearchAbleFields } from "./blog.constant";
 
 const createBlogIntoDB = async (payload: IBlog) => {
-
-    const result = (await Blog.create(payload)).populate("author", "name email role");
-    console.log('kaku', result);
+    const user = await User.findById(payload.author);
 
 
-    // const fetchedBlog = await blogData(blogData._id.toString());
+    if (!user) {
+        throw new Error("Invalid user, author not found");
+    }
 
-    // return result.populate("author", "name email role");
+    const blog = await Blog.create(payload);
+
+    const result = await blog.populate("author", "name email role")
+
     return result;
 }
 
 
-const updateBlogIntoDB = async (id: string, payload: Partial<IBlog>) => {
+const updateBlogIntoDB = async (id: string, payload: Partial<IBlog>, userId: string) => {
 
     const blogData = await Blog.findById(id);
 
-
-
-    // console.log(blogData._id.toString());
-
-
-
-    // console.log(blogData, "blogdata");
-
-    // const userdata = await User.findById(userId);
-    // console.log(userdata);
-
-    // if (!userId) {
-    //     throw new Error("Blog is not found")
-    // }
-
     if (!blogData) {
-        throw new Error("Blog is not found")
+        throw new Error("Blog is not found ! ")
     }
-    const updateBlogData = await Blog.findByIdAndUpdate(id, payload, {
+
+    if (blogData.author.toString() !== userId) {
+        throw new Error("you are not author !")
+    }
+
+    const result = await Blog.findByIdAndUpdate(id, payload, {
         new: true,
         runValidators: true
     })
 
-    const authorData = await User.findOne(blogData?.author);
-
-
-
-    return { updateBlogData, authorData }
+    return result
 };
 
 
 
-const deleteBlogIntoDB = async (id: string) => {
+const deleteBlogIntoDB = async (
+    id: string,
+    userId: string,
+) => {
 
     const blog = await Blog.findById(id);
     if (!blog) {
         throw new Error("Blog is not found")
     }
-    const result = await Blog.findByIdAndDelete(id)
 
-    return result
+    // Verify that the logged In user is the author of the blog
+    if (blog.author.toString() !== userId) {
+        throw new Error("You are not author ! You Can not delete this blog ! ");
+    }
+
+    const result = await Blog.findByIdAndDelete(id);
+
+    return result;
 }
 
 
@@ -75,9 +72,7 @@ const getAllBlogFromDb = async (query: Record<string, unknown>) => {
 
     const result = await blogsQuery.modelQuery;
 
-
-
-    return result
+    return result;
 
 }
